chore(extractor): port the data extractor to Rust

+++ tfhe-benchmark/benches/core_crypto/ks_pbs_bench.rs
@@ -273,7 +273,7 @@ fn ks_pbs<Scalar: UnsignedTorus + CastInto<usize> + Serialize>(
             &benchmark_spec,
             "ks-pbs",
             &OperatorType::Atomic,
-            bit_size,
+            bit_size as u64,
             vec![bit_size],
         );
     }
@@ -506,7 +506,7 @@ fn multi_bit_ks_pbs<
             &benchmark_spec,
             "ks-pbs",
             &OperatorType::Atomic,
-            bit_size,
+            bit_size as u64,
             vec![bit_size],
         );
     }
@@ -834,7 +834,7 @@ mod cuda {
                 &benchmark_spec,
                 "ks-pbs",
                 &OperatorType::Atomic,
-                bit_size,
+                bit_size as u64,
                 vec![bit_size],
             );
         }
@@ -1141,7 +1141,7 @@ mod cuda {
                 &benchmark_spec,
                 "ks-pbs",
                 &OperatorType::Atomic,
-                bit_size,
+                bit_size as u64,
                 vec![bit_size],
             );
         }
